requirements.txt
python-dotenv==1.0.1
langchain
langchain-openai
openai
playwright==1.46.0
beautifulsoup4==4.12.3
tldextract==5.1.2
langdetect==1.0.9
pymupdf==1.24.9
requests==2.32.3
pydantic==2.11.7
tenacity==8.5.0
lxml==5.1.0